Move handle_event to job builder, adjust tests
