+++ lib/lightning_web/live/workflow_live.ex
@@ -5,7 +5,6 @@ defmodule LightningWeb.WorkflowLive do
   on_mount({LightningWeb.Hooks, :project_scope})
 
   alias Lightning.Workflows
-  alias Lightning.Jobs
 
   @impl true
   def render(assigns) do
@@ -13,16 +12,6 @@ defmodule LightningWeb.WorkflowLive do
     <Layout.page_content>
       <:header>
         <Layout.header title={@page_title} socket={@socket}>
-          <.link navigate={
-            Routes.project_job_index_path(@socket, :index, @project.id)
-          }>
-            <div class="inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-secondary-200 hover:bg-secondary-300 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-secondary-500">
-              <div class="h-full">
-                <Heroicons.table_cells solid class="h-4 w-4 inline-block" />
-              </div>
-            </div>
-          </.link>
-          &nbsp;&nbsp;
           <.link navigate={
             Routes.project_workflow_path(@socket, :new_job, @project.id)
           }>
@@ -130,40 +119,6 @@ defmodule LightningWeb.WorkflowLive do
      )}
   end
 
-  @impl true
-  def handle_event("delete", %{"id" => id}, socket) do
-    job = Jobs.get_job!(id)
-
-    case Jobs.delete_job(job) do
-      {:ok, _} ->
-        LightningWeb.Endpoint.broadcast!(
-          "project_space:#{socket.assigns.project.id}",
-          "update",
-          %{}
-        )
-
-        {:noreply,
-         socket
-         |> put_flash(:info, "Job deleted successfully")
-         |> push_patch(
-           to:
-             Routes.project_workflow_path(
-               socket,
-               :show,
-               socket.assigns.project.id
-             )
-         )}
-
-      {:error, _} ->
-        {:noreply,
-         socket
-         |> put_flash(
-           :error,
-           "Unable to delete this job because it has downstream jobs"
-         )}
-    end
-  end
-
   @doc """
   Update the encoded project space, when a change is broadcasted via pubsub
   """
@@ -253,96 +208,5 @@ defmodule LightningWeb.WorkflowLive do
 
     socket
     |> assign(page_title: socket.assigns.project.name, workflow: workflow)
-  end
-
-  @impl true
-  def render(assigns) do
-    ~H"""
-    <Layout.page_content>
-      <:header>
-        <Layout.header title={@page_title} socket={@socket}>
-          <.link navigate={
-            Routes.project_workflow_path(@socket, :new_job, @project.id)
-          }>
-            <Common.button>
-              <div class="h-full">
-                <Heroicons.plus class="h-4 w-4 inline-block" />
-                <span class="inline-block align-middle">New Job</span>
-              </div>
-            </Common.button>
-          </.link>
-        </Layout.header>
-      </:header>
-      <div class="relative h-full">
-        <%= case @live_action do %>
-          <% :new_job -> %>
-            <div class="absolute w-1/3 inset-y-0 right-0 bottom-0 z-10">
-              <div
-                class="w-auto h-full bg-white shadow-xl ring-1 ring-black ring-opacity-5"
-                id="job-pane"
-              >
-                <.live_component
-                  module={LightningWeb.JobLive.JobBuilder}
-                  id="builder-new"
-                  job={@job}
-                  workflow={assigns[:workflow]}
-                  params={@job_params}
-                  project={@project}
-                  current_user={@current_user}
-                  builder_state={@builder_state}
-                  return_to={
-                    Routes.project_workflow_path(@socket, :show, @project.id)
-                  }
-                />
-              </div>
-            </div>
-          <% :edit_job -> %>
-            <div class="absolute w-1/3 inset-y-0 right-0 z-10">
-              <div class="w-auto h-full" id={"job-pane-#{@job.id}"}>
-                <.live_component
-                  module={LightningWeb.JobLive.JobBuilder}
-                  id={"builder-#{@job.id}"}
-                  job={@job}
-                  project={@project}
-                  current_user={@current_user}
-                  builder_state={@builder_state}
-                  return_to={
-                    Routes.project_workflow_path(@socket, :show, @project.id)
-                  }
-                />
-              </div>
-            </div>
-          <% :edit_workflow -> %>
-            <div class="absolute top-0 right-0 m-2 z-10">
-              <div class="w-80 bg-white rounded-md shadow-xl ring-1 ring-black ring-opacity-5 p-3">
-                <.live_component
-                  module={LightningWeb.WorkflowLive.WorkflowInspector}
-                  id={@workflow.id}
-                  workflow={@workflow}
-                  project={@project}
-                  return_to={
-                    Routes.project_workflow_path(
-                      @socket,
-                      :show,
-                      @project.id
-                    )
-                  }
-                />
-              </div>
-            </div>
-          <% _ -> %>
-        <% end %>
-        <div
-          phx-hook="WorkflowDiagram"
-          class="h-full w-full"
-          id={"hook-#{@project.id}"}
-          phx-update="ignore"
-          base-path={Routes.project_workflow_path(@socket, :show, @project.id)}
-          data-project-space={@encoded_project_space}
-        >
-        </div>
-      </div>
-    </Layout.page_content>
-    """
   end
 end
